Test: Isolate core module coverage
- add a TUIkitCore-only test target
- move key, preference, and buffer primitive tests
- retain overlay behavior in the umbrella integration suite
