+++ Tests/TUIkitCoreTests/FrameBufferTests.swift
@@ -1,0 +1,82 @@
+//  🖥️ TUIKit — Terminal UI Kit for Swift
+//  FrameBufferTests.swift
+//
+//  Created by LAYERED.work
+//  License: MIT
+
+import Testing
+
+@testable import TUIkitCore
+
+@MainActor
+@Suite("FrameBuffer Tests")
+struct FrameBufferTests {
+
+    @Test("Empty buffer has zero dimensions")
+    func emptyBuffer() {
+        let buffer = FrameBuffer()
+        #expect(buffer.width == 0)
+        #expect(buffer.height == 0)
+        #expect(buffer.isEmpty)
+    }
+
+    @Test("Single line buffer has correct dimensions")
+    func singleLine() {
+        let buffer = FrameBuffer(text: "Hello")
+        #expect(buffer.width == 5)
+        #expect(buffer.height == 1)
+        #expect(buffer.lines == ["Hello"])
+    }
+
+    @Test("Vertical append stacks lines")
+    func verticalAppend() {
+        var buffer = FrameBuffer(text: "Line 1")
+        buffer.appendVertically(FrameBuffer(text: "Line 2"))
+        #expect(buffer.height == 2)
+        #expect(buffer.lines == ["Line 1", "Line 2"])
+    }
+
+    @Test("Vertical append with spacing")
+    func verticalAppendWithSpacing() {
+        var buffer = FrameBuffer(text: "Top")
+        buffer.appendVertically(FrameBuffer(text: "Bottom"), spacing: 2)
+        #expect(buffer.height == 4)
+        #expect(buffer.lines == ["Top", "", "", "Bottom"])
+    }
+
+    @Test("Horizontal append places side by side")
+    func horizontalAppend() {
+        var buffer = FrameBuffer(text: "Left")
+        buffer.appendHorizontally(FrameBuffer(text: "Right"), spacing: 1)
+        #expect(buffer.height == 1)
+        #expect(buffer.lines == ["Left Right"])
+    }
+
+    @Test("Horizontal append with different heights pads correctly")
+    func horizontalAppendDifferentHeights() {
+        var left = FrameBuffer(lines: ["AB", "CD"])
+        let right = FrameBuffer(text: "X")
+        left.appendHorizontally(right, spacing: 1)
+        #expect(left.height == 2)
+        #expect(left.lines[0] == "AB X")
+        // Row 1: "CD" padded to width 2, spacing " ", no right content
+        #expect(left.lines[1] == "CD ")
+    }
+
+    @Test("ANSI codes are excluded from width calculation")
+    func ansiStrippedWidth() {
+        let styled = "\u{1B}[1mBold\u{1B}[0m"
+        let buffer = FrameBuffer(text: styled)
+        #expect(buffer.width == 4)  // "Bold" is 4 chars
+    }
+
+    @Test("Horizontal append with ANSI codes pads correctly")
+    func horizontalAppendWithAnsi() {
+        let styled = "\u{1B}[1mHi\u{1B}[0m"
+        var left = FrameBuffer(text: styled)
+        left.appendHorizontally(FrameBuffer(text: "There"), spacing: 1)
+        #expect(left.height == 1)
+        // "Hi" (styled) + " " (spacing) + "There"
+        #expect(left.lines[0].stripped == "Hi There")
+    }
+}
