+++ Tests/TUIkitCoreTests/KeyEventTests.swift
@@ -6,7 +6,7 @@
 
 import Testing
 
-@testable import TUIkit
+@testable import TUIkitCore
 
 // MARK: - Key Enum Tests
